setup a grid container on the <body> of 3 columns and 4 rows
the columns width are flexible.

the rows are fixed for the <header> and <nav> and flexible for the last two remaining rows.

a grid gap of 5px is also configured

I've also reordered the CSS selectors to have all direct chidlren of the <body>  just below it are they are the grid items of the grid container associated to the <body> and it make more sense (more readable)