added css  selectors to align images
and a selector to clear the float used for image alignment to avoid images to overlap on other elements

also added margin properties to the image aligment to avoid having them too close to the text.

Added a picture of a blue butterfly to illustrate the overlap issue if float are not cleared.