Made the <main> element a first level container on the same level as header, (site) nav and footer. Inside main there will be sections on the left and aside on the right. On the right there will also be a (page) nav.
That means that there is a second (inner) CSS Grid and <main> is a CSS grid container.

the outer grid is now set to auto-fit one column, while the inner grid use auto-fit/minmax. In theory it should help with responsive design but it doesn't seems to work very well. Wonder if @media queries are still required.

Because there are now two <nav>, one for the site and one internal to the page, I had to adjust the ruleset for styling <ul> and <li> to make them different as it's only the site nav that I want inline and links colors, I want them different too.

Tweeaked the grid-column setting of main grid to use span and show the starting grid line as it convey the information more clearly.