Continuing exploration of responsive images (work in progress, as it still buggy)
Html changes: inspired by a series of articles on responsive images:
https://cloudfour.com/thinks/responsive-images-101-definitions/

I've decided to use the other form (using srcset and sizes on <img>, which is more appropriate for resolution switching) instead of using <picture> (which is more for art direction where the picture source is slightly different but similar enough so the alt attributes on the still mandatory <img> still describes the different sources inside <picture>).

srcset and sizes on <img> are more finicky to configure (coming up with image breakpoint is a pain) and getting it to work across browsers is still hard.
At the moment, only Chrome is loading the right image for each sizes string.
Both Firefox and Safari are loading the image from the next breakpoint up for under 400px viewport size.

Css changes: only float the second image to the right of the paragraph of text if the screen is wide enough. On small screen use full width (inspired by the walmart example in the article above)