So if using srset and sizes with <img>, then the width attribute on <img> doesn't make sense. Removing it improved the responsive rendering a bit as the space the image occupies on the screen matches what's described in sizes attributes.
adding a viewport meta-tag improve the cross-browser discrepancies, so now Safari and Chrome are displaying images as intended. I learnt about he viewport meta-tag here:
https://developer.mozilla.org/en-US/docs/Mozilla/Mobile/Viewport_meta_tag

The choice of which of the images in the srcet is left to the browsers to decide.

There might still be an issue for viewport between. 320px (breakpoint for the second image) and 550px (breakpoint for the layout, including when the second image become a right float)